rule induction(variable, expr)
Invoked on (v, e), this attempts to prove relation A(v) for all v >= e.
This requires that the last relation derived is of the form
A(v) -> A(v + 1), and that the previous relation derived is
of the form A(e).

For now, this just claims the new relation A(v); we'll need set
handling to attach the v >= e condition to it in the future.