More useful expression diff
diff() now compares expressions as is first, then compares clean versions.
The location of difference on the as-is versions is returned if they
remain different.

This removes almost all need for anything else to invoke clean().

Failure to validate a line now shows both the unmatched portions of
the uncleaned expressions, and the whole of the cleaned versions.