rule itervar to change variable on an iterator
This allows sums and products to be rewritten replacing eg i with i + 1
or with n - i (since they both preserve a step size of 1).

For now the replacement cannot reference a bound variable, even one
within scope of the iterator; so for example in \sum_i=...{sum_j=...}
we cannot replace j with i-j.