Make derivation rules quantifier-safe
Note: we no longer resolve the expression to derive directly after parsing,
since the derivation may itself modify the context in which it should be
resolved; instead we do this at the point of validation.

Rule changes:
- new signature identity(varlist, expr): the listed variables may be
referred to in the expression; the result will be of the form:
  \Aa: \Ab: ... expr = expr
- new signature condstart(varlist): the listed variables are introduced
as free variables for the scope of the conditional proof.
- new signature condend(varmap): the listed variables, which should have
been free for the scope of the conditional proof, are mapped to new
quantified (universal) variables.
- note also that Axiom::Context now supports a modified dictionary within
the scope of the conditional proof.
- induction is now more rigorous: we now prove \Ax: f(x), given antecedents
f(base) and \Ax: f(x) -> f(x+1).
- equate(), add(), multiply() and recurse() now allow for one or more
quantifiers to wrap the primary 'equals' operator.
- lots of resolving changes to resolve supplied sub-expressions in the
correct context.