Prove more sum identities
This adds Sum_pow_2_i and Sum_pow_m2_i.

It also demonstrates a more constructive approach to finding sum
identities: whereas with induction you need to know the destination
in advance, extending and then rebasing the sum in many (and predictable)
cases allows the result to be discovered in the process.