wip minimize testcase: fail
This change - no longer blessing an arrayref that is already being used
only as an arrayref - appears to be enough to make the testcase
deterministic.

This implies that something in the act of blessing an arrayref, or of
creating a new object that will later have DESTROY called on it, is
somehow affecting the determinism of hash operations.