Improved mullist->clean, pow->clean
Replace specific handling to elide x(a, recip(a)) with a more general
approach that collates powers of the same expression.

This replaces distribution of pow(a, b+c) in clean, but we add instead
pow(c1, c2 + a) -> x(eval(c1^c2), pow(c1, a)) and separation of integer
powers of -1 for negated values.