Bugfix: iterextend could match wrong iterators
Given sumA + b = sumA + c, we derive by first excluding any iterators
that are identical in the before and after expressions, expecting
that what remains will point us to the one to extend. However we did
not restrict matches to being on the same side of the '=' sign, so
if we attempted to derive sum{A+1} + b' = sumA + c we'd pick the
wrong one for trial extend, and fail.

Fixed this to match up only before/after iterators on the same side
of the equals sign.