bugfix: extend diff() to take an 'exact' option
When cleaning, the collation of powers in a mullist, or of like terms
in a pluslist, uses diff() to check if two subexpressions of the same
Expr are the same. In this context we must treat [local i_1] and
[local i_2] as distinct - they're resolved as part of the same
expression - but we want to treat [sum i_1 ...] and [sum i_2 ...]
as the same if mapping i_1 to i_2 would find them so.

The new 'exact' option does this, and gets used by the two clean()
rules. It should in future be used anywhere else we're calling diff()
on two subexpressions of the same resolved Expr.