Workaround for 'recurse' crashes
If we find a matching subexpression in the target, but it involves
variables local to a small part of the target, we get a resolve failure
in the count.

For the moment, catch that with an eval so we can report that this is
not valid; better, though, would be to avoid including these in the
result stack to try.