Add support for integration
New expr \int_{x=a}^{b}{f}, of exactly the same form as \sum, representing
the integral for x=a to x=b of f dx.
New expr \inteval_{x=a}^{b}{g}, of exactly the same form as \sum,
representing the evaluation of the same integral, where g' = f.
New derivation itersplit, to split an integral (or other iterator) in two.
New derivation integrate, to replace \int with \inteval.
New derivation inteval, to replace \inteval with its evaluation.

TODO: add support for inequalities to tighten this up
TODO: add helper function to reduce duplication of code across derivations
like IterExpand and IntEval that expect the diff to point midway into an
expression on the RHS.