Extend derivation multiply() to support a value expression
Use eg 'multiply(*value a+b)' to specify the multiplicand rather than
leaving it to derive() to determine it.

TODO: this was added as a general-purpose stopgap after seeing that
we failed to determine the multiplicand when trying to multiply
'a+b <= c+d' by -1. But we should also improve the discovery code.