Extend derivation itersplit() to check range validity
For an iterator for var a from b to c that we are splitting on d,
we now verify that b <= d <= c. Where d-b and c-d are constant
that is done directly; where not, we use a combination of ranges
specified for enclosing iterators and an optional 'with' statement
for the verification.

Thus for example:

theorem C1: \Aa: ((a >= 0) & (a <= 1)) -> ((1-a >= 0) & (1-a <= 1))
axiom A1: x = \int_{a=0}^1{\int_{b=0}^1{y}}
itersplit(*with C1): x = \int_{a=0}^1{\int_{b=0}^{1-a}{y} + \int_{b=1-a}^1{y}}

Much of the heavy lifting is done by the new Axiom::Derive->check_range():
  $self->check_range($targ, $base, $loc, $with)
.. which verifies that $targ->isa('Axiom::Expr') is true, given iterator
ranges from the $base->isa('Axiom::Expr') enclosing the location $loc,
and the (optional) named theorem $with.