Tweak grammar for correctness (statements) and efficiency
Rearrange Statement/SStatement for correctness, and (slightly) for
efficiency. Tweak some ambiguous tokens for efficiency.

Allow a variable to represent a proposition. This is dubious: it implies
supporting second-order logic, and it is not clear if we are ready for that.

For efficiency we'd also like a) to factor out common prefixes such
as <[args=SStatement]>, and b) to disallow backtracking over unambiguous
tokens, where possible.