New derivation rename()
Given a prior theorem of the form P(x), constructs the new theorem P(y).

FIXME: this introduces a tweak to diff() to cope with a new type of
latitude, but it is far from clear what the existing variants do and
whether this tweak is needed. To investigate, document, test and maybe
rationalize.