Add Axiom::Expr->unencumbered
A boolean expression A is unencumbered if when A -> B we can safely
replace A with B. In particular, in the expression A -> C, A is _not_
unencumbered.

This should be used both by the 'ponens' derivation and by other
derivations such as 'multiply' which implement an implicit implication
(since A = B -> cA = cB, but this is not reversible when c can be zero).