Rewrite derivation parsing to extend naming
Derivations are now always of the form:
  <derivation name> [ ( args, ... ) ]? <theorem name>?: <expr>

Individual derivations need only supply the pattern for the optional
list of arguments (without the parens), and the count of arguments
(so that the initialization can happen automatically).

'axiom' and 'theorem' derivations explicitly mark their names as
exportable. This is just a placeholder for now, but the intent is
that only file-scope exportable names will be exported under '*import'.