Extend derivation 'specify'
Lift the 'only one variable' restriction; document additional restrictions;
verify that each specified expression is numeric over its range.

TODO:
- check for shadowing of free variables;
- determine final disposition and naming of helper class Axiom::ExprVars;

Ideally we would find a way to abstract the hand-rolled walk-and-resolve
process to reduce or remove the knowledge of ::Expr internals. But that
is made hard by the fact that specified expressions may reference local
variables: we need to be halfway through the walk-and-resolve before we
even discover what they mean.