Add $expr->apply_map()
We walk the expr to apply the replacements, resolving as we go. Variables
getting introduced are resolved but not replaced; others are replaced if
their old resolution matches a map id, else get resolved normally.
Replacement exprs are resolved but not recursively replaced, and we verify
that each replacement for a given id resolves the same way as the first.
We also check that each replacement has a numeric value across its range.

TODO: check current uses of $expr->subst_var{,s} to see if they should
be replaced by a call to this method.