Rewrite derivation unarydistrib
All distributions now support grouping, so that f(a+b+c) can be
transformed not only to f(a)+f(b)+f(c), but also to f(a+c)+f(b).
Add tests of supported forms.

As a result, derivation now does a lot more work, and validation
has become purely mechanical.

Support for 'prod' is removed for now, will be added back in if/when
we need it.